requirements.txt
fastapi
authlib
pymongo
httpx
python-dotenv
PyJWT
itsdangerous
pydantic
starlette
pyYAML
uvicorn
cryptography
h11
sniffio
certifi
click
dnspython
rsa
cryptography
python-multipart